Moved the downloaders to Downloaders folder and made the private
