+++ Source/RpmAvatarCreator/Private/Downloaders/RpmAuthManager.cpp
