+++ Source/RpmAvatarCreator/Private/RpmAvatarCreatorApi.cpp
@@ -4,12 +4,12 @@
 #include "RpmAvatarCreatorApi.h"
 
 #include "ReadyPlayerMeSettings.h"
-#include "RpmPartnerAssetDownloader.h"
-#include "RpmColorDownloader.h"
-#include "RpmAuthManager.h"
+#include "Downloaders/RpmPartnerAssetDownloader.h"
+#include "Downloaders/RpmColorDownloader.h"
+#include "Downloaders/RpmAuthManager.h"
 #include "RpmAvatarRequestHandler.h"
 #include "Requests/RequestFactory.h"
-#include "RpmUserAvatarDownloader.h"
+#include "Downloaders/RpmUserAvatarDownloader.h"
 #include "RpmAvatarTemplateDownloader.h"
 #include "RpmImageDownloader.h"
 #include "Engine/SkeletalMesh.h"
